# perf: more scalable DirectTransmission batching