## Which problem is this PR solving?
In the original implementation, goroutine and timer proliferation were a concern.

## Short description of the changes
Note: based on #1604

This replaces a whole bunch of channel and goroutine contention with a whole bunch of lock contention. I think that's going to be ok because the stuff we do while holding the lock is very cheap, and ultimately less than what libhoney was doing. It bechmarks well - the remaining allocation here is mostly in the http transport (including the test server itself) and there's not much to be done about that.

```
benchmark                                               old ns/op     new ns/op     delta
BenchmarkTransmissionComparison/high_card/direct-12     1101          702           -36.23%
BenchmarkTransmissionComparison/low_card/direct-12      894           640           -28.39%

benchmark                                               old allocs     new allocs     delta
BenchmarkTransmissionComparison/high_card/direct-12     3              2              -33.33%
BenchmarkTransmissionComparison/low_card/direct-12      3              2              -33.33%

benchmark                                               old bytes     new bytes     delta
BenchmarkTransmissionComparison/high_card/direct-12     5035          3050          -39.42%
BenchmarkTransmissionComparison/low_card/direct-12      4864          3211          -33.98%
```